# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi API untuk sistem pencatatan target dan performa Karyawan (Employee Tracker).",
        "title": "Employee Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/categories/{category_id}/products": {
            "get": {
                "description": "Mengambil daftar produk berdasarkan ID Kategorinya",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "List Produk Berdasarkan Kategori",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kategori",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil produk",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-array_github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_ProductWithCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Format UUID salah",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/products": {
            "get": {
                "description": "Mengambil daftar semua produk beserta detail kategorinya",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "List Semua Produk",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data produk",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-array_github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_ProductWithCategoryResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan data produk baru di bawah kategori tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Buat Produk Baru",
                "parameters": [
                    {
                        "description": "Payload data produk",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Produk berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_ProductWithCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Format JSON salah atau validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "description": "Menampilkan detail informasi produk berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Detail Produk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Produk",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail produk",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_ProductWithCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Format UUID salah",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Produk tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbaharui data produk yang ada di dalam sistem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update Data Produk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Produk",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload data produk",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk berhasil diperbaharui",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_ProductWithCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Format JSON salah atau validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Produk tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_norm": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.ProductRequest": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "category_id": {
                    "type": "string",
                    "example": "uuid"
                },
                "name": {
                    "type": "string",
                    "example": "KUR Bank"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.ProductWithCategoryResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.CategoryResponse"
                },
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_norm": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "example": "pesan error"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-array_github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_ProductWithCategoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.ProductWithCategoryResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_ProductWithCategoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.ProductWithCategoryResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}